Read hostnames to check from a .dropbox-cam-hosts file
Instead of hard coding the hostnames in the script, read hostnames from
a /home/pi/.dropbox-cam-hosts file. This file should be one hostname per
line, no whitespace. The script pings each hostname in the file. If the
ping is successful (ie the pinged device is on the network) the script
exits without capturing or uploading an image.